task-2a/requirements.txt
absl-py==1.4.0
astunparse==1.6.3
audioread==3.0.0
blosc2==2.0.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.33
datetime-glob==1.0.8
decorator==5.1.1
flatbuffers==23.3.3
fonttools==4.39.2
gast==0.5.3
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
idna==3.4
imageio==2.26.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
joblib==1.2.0
keras==2.8.0
Keras-Preprocessing==1.1.2
ketos==2.6.2
kiwisolver==1.4.4
lazy_loader==0.1
lexery==1.1.2
libclang==15.0.6.1
librosa==0.10.0.post1
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
msgpack==1.0.5
networkx==3.0
numba==0.56.4
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
Pint==0.20.1
platformdirs==3.1.1
pooch==1.7.0
protobuf==3.20.3
psutil==5.9.4
py-cpuinfo==9.0.0
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.4
tables==3.8.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tifffile==2023.3.15
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
version-parser==1.0.1
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
